RavenDB-11310 moving files around
