+++ src/Raven.Server/Documents/Handlers/BatchHandler.cs
@@ -26,6 +26,7 @@ using Raven.Server.Documents.Patch;
 using Raven.Server.Documents.PeriodicBackup;
 using Raven.Server.Documents.Replication;
 using Raven.Server.Documents.TimeSeries;
+using Raven.Server.Documents.TransactionMerger;
 using Raven.Server.Json;
 using Raven.Server.Rachis;
 using Raven.Server.Routing;
