+++ src/Raven.Server/Documents/Handlers/BatchRequestParser.cs
@@ -18,9 +18,8 @@ using Raven.Client.Documents.Session;
 using Raven.Client.Documents.Smuggler;
 using Raven.Client.Util;
 using Raven.Server.Documents.Patch;
-using Raven.Server.Documents.PeriodicBackup;
-using Raven.Server.Documents.TransactionCommands;
-using Raven.Server.Exceptions;
+using Raven.Server.Documents.TransactionMerger.Commands;
+using Raven.Server.Documents.PeriodicBackup;using Raven.Server.Exceptions;
 using Raven.Server.ServerWide;
 using Sparrow;
 using Sparrow.Json;
