+++ src/Raven.Server/Documents/Handlers/HiLoHandler.cs
@@ -12,6 +12,7 @@ using System.Threading.Tasks;
 using Raven.Client;
 using Raven.Client.Documents.Changes;
 using Raven.Client.Exceptions.Documents;
+using Raven.Server.Documents.TransactionMerger;
 using Raven.Server.Routing;
 using Raven.Server.ServerWide.Context;
 using Raven.Server.TrafficWatch;
