+++ src/Raven.Server/Documents/Replication/OutgoingReplicationHandler.cs
@@ -39,7 +39,7 @@ using Sparrow.Server;
 using Sparrow.Server.Utils;
 using Sparrow.Threading;
 using Sparrow.Utils;
-using Voron;
+using Raven.Server.Documents.TransactionMerger;
 
 namespace Raven.Server.Documents.Replication
 {
